Beautification of the moment view
Add cards and ionic button styles

Add concatCss and allowed Sass transpiling

Remove dynamic text are plugin as it was breaking tests

Removed dynamic textarea plugin

Adds new styles to moment-view

Fixes typo and several formatting issues

Removes angular-elastic from bower.json

This removes a third party angular plugin that allowed easy resize
of a text area given the amount of text in it. It was removed because
it was breaking the tests.

We might be able to add it later when we're in the polishing stages.

Restores $httpProvider.defaults.withCredentials = true;